requirements.txt
PydoNovosoft==1.3.1
requests
json_logging==0.0.10
jinja2
psycopg2-binary
fpdf
pandas
logging
flask